Finished User SignUp Use Case

+++ src/main/java/entity/Activity.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 // Public class Activity that inherits from Business and implements Timeblock
 public class Activity extends Business implements Timeblock {
-    public Activity(String name, ArrayList<Float> location, float distance, String contactNum,
+    public Activity(String name, ArrayList<Double> location, Double distance, String contactNum,
                     String price, Float rating) {
         super(name, location, distance, contactNum, price, rating);
     }
